refactor(modules): extract prototypes/Object.mjs from prototype.mjs
