+++ features/history/history.js
@@ -1,4 +1,4 @@
-import '../../modules/prototype.mjs'
+import '../../modules/prototypes/Object.mjs'
 import '../../modules/prototypes/localStorage.mjs'
 import * as readingList from '../reading-list/readingList.mjs'
 import * as statusBar from '../status-bar/statusBar.js'
